% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adresser.R
\name{adresser}
\alias{adresser}
\title{Adressesøgning}
\usage{
adresser(...)
}
\arguments{
\item{...}{Find parametre her:
\href{https://dawa.aws.dk/adressedok#adressesoegning}{Adresse parametre}}
}
\value{
En adresse er en struktureret betegnelse som angiver en særskilt adgang
til et areal, en bygning eller en del af en bygning efter reglerne i
adressebekendtgørelsen. Du kan læse mere om de værdier funktionen returnerer
her: \href{https://dawa.aws.dk/adressedok#adressedata}{DAWA adressedata}
}
\description{
Søg efter adresser. Returnerer de adresser som opfylder kriteriet. Med mindre
der er behov for felter som kun er med i den fulde adressestruktur anbefaler
vi at, man tilføjer parameteren struktur=mini, da dette vil resultere i bedre
performance.
}
\examples{
\dontrun{
# Find de adresser som ligger på Rødkildevej og har husnummeret 46.
dawa::adresser(vejnavn = "Rødkildevej", husnr = 46, struktur = "mini")

# Find de adresser som ligger på Rødkildevej og har husnummeret 46.
# Resultatet leveres i geojson format.
dawa::adresser(vejnavn = "Rødkildevej", husnr = 46, format = "geojson", struktur = "mini")

# Find de adresser som ligger på Rødkildevej og har husnummeret 46.
# Resultatet leveres i csv format.
dawa::adresser(vejnavn = "Rødkildevej", husnr = 46, format = "csv")

# Find de adresser som indeholder et ord der starter med hvid og har postnummeret 2400
dawa::adresser(q = "hvid*", postnr = 2400)

# Find de adresser som er indenfor polygonet (10.3,55.3), (10.4,55.3),
# (10.4,55.31), (10.4,55.31), (10.3,55.3)
dawa::adresser(polygon = "[[[10.3,55.3],[10.4,55.3],[10.4,55.31],[10.4,55.31],[10.3,55.3]]]")

# Find den adresse, som har KVHX-nøgle 04615319__93__1____
dawa::adresser(kvhx = "04615319__93__1____")

# Hent alle adresse i postnummer 8471, i GeoJSON format, med koordinater
# angivet i ETRS89 / UTM zone 32N (SRID 25832)
dawa::adresser(postnr = 8471, format = "geojson", srid = 25832)
}
}
