% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.R
\name{reverse}
\alias{reverse}
\title{Reverse geokodning
Find den adgangsadresse, som ligger nærmest det angivne koordinat. Som koordinatsystem kan anvendes ETRS89/UTM32 med srid=25832 eller WGS84/geografisk med srid=4326. Default er WGS84.}
\usage{
reverse(...)
}
\arguments{
\item{...}{En af de nedenstående parametre.}

\item{x}{X koordinat. (Hvis ETRS89/UTM32 anvendes angives øst-værdien.) Hvis WGS84/geografisk anvendex angives bredde-værdien.}

\item{y}{Y koordinat. (Hvis ETRS89/UTM32 anvendes angives nord-værdien.) Hvis WGS84/geografisk anvendex angives længde-værdien.}

\item{srid}{Angiver SRID for det koordinatsystem, som geospatiale parametre er angivet i. Default er 4326 (WGS84).}

\item{callback}{Output leveres i JSONP format. Se Dataformater.}

\item{format}{Output leveres i andet format end JSON. Se Dataformater.}

\item{noformat}{Parameteren angiver, at whitespaceformatering skal udelades}

\item{struktur}{Angiver om der ønskes en fuld svarstruktur (nestet), en flad svarstruktur (flad) eller en reduceret svarstruktur (mini).  Mulige værdier: "nestet", "flad" eller "mini".  For JSON er default "nestet", og for CSV og GeoJSON er default "flad". Det anbefales at benytte mini-formatet hvis der ikke er behov for den fulde struktur, da dette vil give bedre svartider.}
}
\value{
En adgangsadresse er en struktureret betegnelse som angiver en særskilt adgang til et areal eller en bygning efter reglerne i adressebekendtgørelsen. Forskellen på en adresse og en adgangsadresse er at adressen rummer eventuel etage- og/eller dørbetegnelse. Det gør adgangsadressen ikke. Du kan læse mere om de værdier funktionen returnerer her: \href{https://dawa.aws.dk/adgangsadressedok#adressedata}{DAWA adgangsadressedata}
}
\description{
Reverse geokodning
Find den adgangsadresse, som ligger nærmest det angivne koordinat. Som koordinatsystem kan anvendes ETRS89/UTM32 med srid=25832 eller WGS84/geografisk med srid=4326. Default er WGS84.
}
\examples{
\dontrun{
# Returner adgangsadressen nærmest punktet angivet af WGS84/geografisk koordinatet (12.5851471984198, 55.6832383751223)
dawa::reverse(x = 12.5851471984198, y = 55.6832383751223, struktur = mini)

# Returner adressen nærmest punktet angivet af ETRS89/UTM32 koordinatet (725369.59, 6176652.55)
dawa::reverse(x = 725369.59, y = 6176652.55, srid = 25832)
}
}
